split up store.ts file for better readability
